refactor(lsp): split document-links into dedicated microcrate
