+++ crates/perl-lsp-navigation/Cargo.toml
@@ -25,7 +25,7 @@ serde_json = "1.0.149"
 url = "2.5.8"
 perl-position-tracking = { workspace = true }
 perl-module-path = { workspace = true }
-perl-module-import = { workspace = true }
+perl-lsp-document-links = { workspace = true }
 
 [features]
 default = []
